+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - App config for SQL
+- App config for Airbrake
 
 
 ## [0.6.3] - 2017-12-05